requirements.txt
attrs==19.3.0
requests
beautifulsoup4
requests-html